Backend file restructure and readme update
