+++ nodestudio/process/core/mask.py
